% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreAnalysis.R
\name{AnnotateScoreDistribution}
\alias{AnnotateScoreDistribution}
\title{AnnotateScoreDistribution}
\usage{
AnnotateScoreDistribution(x, plot = NULL, ...)
}
\arguments{
\item{x}{an object of class signatureTester}
}
\value{
a ggplot object
}
\description{
Function to add summaries to the density plots
}
\details{
This function annotates a summary plot
}
\examples{
AnnotateScoreDistribution(x, plot)
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
