% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureGeneration.R
\name{calculateMeanExpression}
\alias{calculateMeanExpression}
\title{calculateMeanExpression}
\usage{
calculateMeanExpression(x, Signatures = NULL)
}
\arguments{
\item{x}{an object of class signatureTester}

\item{signatures}{a vector of different gene signatures to be tested (default NULL)}
}
\value{
This a copy of the class with the mean expression of each signature in the scores slot
}
\description{
Function for calculating a "mean expression" score
}
\details{
This function will calculate the mean expression of the genes in each signature and store the resultant dataframe in the scores slot of the class
}
\examples{
calculateMeanExpression(x = object)
calculateMeanExpression(x = object, sigs = c("sig1", "sig2"))
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
