% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreAnalysis.R
\name{combineGroupsScores}
\alias{combineGroupsScores}
\title{combineGroupsScores}
\usage{
combineGroupsScores(x)
}
\arguments{
\item{x}{an object of class signatureTester}
}
\value{
a dataframe of the scores and groups combined.
}
\description{
Function for combining the scores and the assigned groups
}
\details{
This function grabs the scores and the groups and creates a dataset so they can be compared.
}
\examples{
combineGroupsScores(x = object)
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
