% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreAnalysis.R
\name{combinePhenoGroup}
\alias{combinePhenoGroup}
\title{combinePhenoGroup}
\usage{
combinePhenoGroup(x)
}
\arguments{
\item{x}{an object of class signatureTester}
}
\value{
a dataframe of the phenodata from the eSet and the signature groups
}
\description{
Function for returning pheno data from the eSet along with the assigned groups
}
\details{
This function compares the scores and the groups
}
\examples{
combinePhenoGroup(x = object)
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
