% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreAnalysis.R
\name{compareGroupScores}
\alias{compareGroupScores}
\title{compareGroupScores}
\usage{
compareGroupScores(x)
}
\arguments{
\item{x}{an object of class signatureTester}
}
\value{
a boxplot of the scores across the groups in each signature
}
\description{
Function for comparing the scores and the assigned groups
}
\details{
This function compares the scores and the groups
}
\examples{
compareGroupScores(x = object)
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
