% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{filterSignatureGenes}
\alias{filterSignatureGenes}
\title{filterSignatureGenes}
\usage{
filterSignatureGenes(
  x,
  missingGenes = "remove",
  Signatures = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{an object of class signatureTester}

\item{missingGenes}{what to do with the missing genes (currently "remove" or nothing)}

\item{Signatures}{a vector of different gene signatures to be tested (default NULL)}

\item{verbose}{will output some metrics about the genes}
}
\value{
This returns a list of the signatures will the genes not found in the stored eSet removed.
}
\description{
Function for filtering genes out of the signatures
}
\details{
This function returns the signatures requested (default: all) and will also filter out the genes that are not in the stored eSet
}
\examples{
filterSignatureGenes(x = object)
filterSignatureGenes(x = object, missingGenes = "remove", sigs = c("sig1", "sig2"))
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
