% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreAnalysis.R
\name{GetPlotDimensions}
\alias{GetPlotDimensions}
\title{GetPlotDimensions}
\usage{
GetPlotDimensions(plot, ...)
}
\arguments{
\item{plot}{a ggplot object}
}
\value{
a ggplot object
}
\description{
Function to extract the max x and y from a plot
}
\details{
This function plots the scores
}
\examples{
GetPlotDimensions(plot)
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
