% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{hardFilterSignatures}
\alias{hardFilterSignatures}
\title{hardFilterSignatures}
\usage{
hardFilterSignatures(
  x,
  missingGenes = "remove",
  Signatures = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{an object of class signatureTester}

\item{missingGenes}{what to do with the missing genes (currently "remove" or nothing)}

\item{Signatures}{a vector of different gene signatures to be tested (default NULL)}

\item{verbose}{will output some metrics about the genes}
}
\value{
This returns a list of the signatures will the genes not found in the stored expressionSet removed.
}
\description{
Function for removing missing genes from the signatures permanently
}
\details{
This function returns a new copy of the class with the signatures filtered to only the genes that exist in the eSet

This can be useful if you only want to store the genes that exist, but by default the downstream functions will soft filter these genes too.
}
\examples{
hardFilterSignatures(x = object)
hardFilterSignatures(x = object, missingGenes = "remove", sigs = c("sig1", "sig2"))
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
