% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreAnalysis.R
\name{plotScoreDistribution}
\alias{plotScoreDistribution}
\title{plotScoreDistribution}
\usage{
plotScoreDistribution(x, ...)
}
\arguments{
\item{x}{an object of class signatureTester}
}
\value{
a ggplot object
}
\description{
Function for plotting the density of the scores
}
\details{
This function plots the scores
}
\examples{
plotScoreDistribution(signatureClass)
plotScoreDistribution(signatureClass, Signatures = "Signature1")
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
