% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreAnalysis.R
\name{returnPhenoColumns}
\alias{returnPhenoColumns}
\title{compareGroupScores}
\usage{
returnPhenoColumns(x, Columns = NULL)
}
\arguments{
\item{x}{an object of class signatureTester}

\item{columns}{a vector of the columns to return (default = NULL)}
}
\value{
a boxplot of the scores across the groups in each signature
}
\description{
Function for returning columns from the expression set phenotype data
}
\details{
This function compares the scores and the groups
}
\examples{
returnPhenoColumns(x = object)
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
