% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreAnalysis.R
\name{returnScores}
\alias{returnScores}
\title{returnScores}
\usage{
returnScores(x, Signatures = NULL, ...)
}
\arguments{
\item{x}{an object of class signatureTester}

\item{Signatures}{a vector of the signatures to return (default = NULL)}
}
\value{
a dataframe of the scores for the chosen signatures
}
\description{
Function for returning the scores in the class
}
\details{
This function returns the scores
}
\examples{
returnScores(signatureClass)
returnScores(signatureClass, Signatures = "Signature1")
returnScores(signatureClass, Signatures = c("Signature1", "Signature2"))
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
