% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreAnalysis.R
\name{returnSigGeneExpression}
\alias{returnSigGeneExpression}
\title{returnSigGeneExpression}
\usage{
returnSigGeneExpression(x, Signatures = NULL)
}
\arguments{
\item{x}{an object of class signatureTester}

\item{signatures}{a vector of the signatures to return (default = NULL)}
}
\value{
The gene expression values for the genes in the specified signatures
}
\description{
Function for returning the gene expression of genes in the specified signatures
}
\details{
This function compares the scores and the groups
}
\examples{
returnSigGeneExpression(x = object)
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
