% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreAnalysis.R
\name{returnSignatureGenes}
\alias{returnSignatureGenes}
\title{returnSignatureGenes}
\usage{
returnSignatureGenes(x, Signatures = NULL)
}
\arguments{
\item{x}{an object of class signatureTester}

\item{signatures}{a vector of the signatures to return (default = NULL)}
}
\value{
a boxplot of the scores across the groups in each signature
}
\description{
Function for returning the genes in the specified signatures
}
\details{
This function compares the scores and the groups
}
\examples{
returnSignatureGenes(x = object)
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
