% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{returnSignatures}
\alias{returnSignatures}
\title{returnSignatures}
\usage{
returnSignatures(x, Signatures = NULL)
}
\arguments{
\item{x}{an object of class signatureTester}

\item{signatures}{a vector of different gene signatures to be tested (default NULL)}
}
\value{
This returns a list of the signatures requested
}
\description{
Function for returning the signatures
}
\details{
This function returns the genes in the signatures slot of the class. Optionally filtered by signature name
}
\examples{
returnSignatures(x = object)
returnSignatures(x = object, sigs = c("sig1", "sig2"))
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
