% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreAnalysis.R
\name{scaleScores}
\alias{scaleScores}
\title{scaleScores}
\usage{
scaleScores(x)
}
\arguments{
\item{x}{an object of class signatureTester}
}
\value{
This returns the class with the scores slot z scored
}
\description{
Function for zscoring scores
}
\details{
This function will zscore the individual signature scores and return a new copy of the class.
}
\examples{
scaleScores(x = object)
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
