% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureTesterClass.R
\docType{class}
\name{signatureTester-class}
\alias{signatureTester-class}
\alias{signatureTester}
\title{signatureTester}
\description{
Class for testing expression signatures
}
\details{
This class will hold your segments and sample information
}
\section{Slots}{

\describe{
\item{\code{rawExpression}}{this will store the initial expression set that is built when the class is created.}

\item{\code{signatures}}{this is a list of the signatures to be tested.}
}}

\examples{
signatureTester()
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
