% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureTesterClass.R
\name{signatureTesterFromExpressionSet}
\alias{signatureTesterFromExpressionSet}
\title{signatureTesterFromExpressionSet}
\usage{
signatureTesterFromExpressionSet(eSet, sigs = NULL)
}
\arguments{
\item{eSet}{this is an expressionset containing the gene expression of the data you are testing the signature in}

\item{signatures}{a list of different gene signatures to be tested (default NULL)}
}
\description{
Function for building the class from an eSet
}
\details{
This function will generate an instance of class "signatureTester"
}
\examples{
signatureTesterFromExpressionSet(eSet = eSet)
signatureTesterFromExpressionSet(eSet = eSet, sigs = Signatures)
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
