% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreAnalysis.R
\name{SummariseScores}
\alias{SummariseScores}
\title{SummariseScores}
\usage{
SummariseScores(x, ...)
}
\arguments{
\item{x}{an object of class signatureTester}
}
\value{
a dataframe of the score summaries
}
\description{
Function for returning the scores in the class
}
\details{
This function returns the scores
}
\examples{
SummariseScores(signatureClass)
SummariseScores(signatureClass, Signatures = "Signature1")
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
