% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreAnalysis.R
\name{thresholdScores}
\alias{thresholdScores}
\title{thresholdScores}
\usage{
thresholdScores(x, method = "ntile", n = 2, thres = NULL)
}
\arguments{
\item{x}{an object of class signatureTester}

\item{method}{the method for thresholding (current options: "ntile")}

\item{n}{the number of groups, if relevant (default: 2)}
}
\value{
This returns the class with groups added to the assignment slot
}
\description{
Function for thresholding scores
}
\details{
This function offers a few ways of thresholding the scores found in the scores slot of the class
}
\examples{
thresholdScores(x = object, method = "ntile", n = 2)
}
\keyword{expression}
\keyword{signatures}
\keyword{testing}
